requirements.txt
fastapi==0.115.12
openai-agents==0.0.16
uvicorn==0.24.0 
python-dotenv==1.1.0
google-generativeai==0.8.3
Pillow==10.4.0
replicate==0.25.1
